Fix bug (Use WebDevIconsUnicodeDecorate*)
