+++ autoload/leaderf/python/leaderf/mruExpl.py
@@ -9,6 +9,7 @@ from .utils import *
 from .explorer import *
 from .manager import *
 from .mru import *
+from .devicons import getWebDevIconsGetFileTypeSymbol, removeDevIcons
 
 
 #*****************************************************
