+++ autoload/leaderf/python/leaderf/utils.py
@@ -159,18 +159,3 @@ def lfActualLineCount(buffer, start, end, col_width):
     for i in buffer[start:end]:
         num += (int(lfEval("strdisplaywidth('%s')" % escQuote(i))) + col_width - 1) // col_width
     return num
-
-def getWebDevIconsGetFileTypeSymbol(file, isdir=False):
-    isdir = 1 if isdir else 0
-    return lfEval('WebDevIconsGetFileTypeSymbol("{}", {})'.format(file, isdir))
-
-def removeDevIcons(func):
-    @wraps(func)
-    def deco(*args, **kwargs):
-        if lfEval("get(g:, 'Lf_ShowDevIcons', 0)") == "1":
-            _args = list(args)
-            line = _args[1]
-            _args[1] = line[line.find(' ')+1:]
-            args = tuple(_args)
-        return func(*args, **kwargs)
-    return deco
